feat: implement Autocomplete component with debounce and user selection
